Improved refresh token support
Move refresh token logic from the OAuth2CodeGrant class to the base OAuth2 class, so that it applies whenever the server returns a refresh token (for any grant type).

Support the tokenURL parameter for all grant types, and use it when requesting a JSON response. If it's nil, fall back to using the authURL instead.

Fix up some comments.