Add workaround for issue in Safari
There's an issue with Safari, references in some websites, where you cant open
the url again if you press 'Cancel' once when asking to go back to original
app. For the user to workaround this issue it has to restart Safari, which is
not really nice. This fix instead gives the option to append a parameter to
the url, which will make it unique, thus the browser till prompt the user again
next time the app authenticates.