Added a guard check to request the shared UIApplication instance usin… (#212)
Added a guard check to request the shared UIApplication instance using the "key for value" method on UIApplication to bypass the compiler warnings allowing OAuth2 library to be used within a library that is used as a private framework for an iOS app that has app extensions. This performs an optionality check on UIApplication, so that if its available to be called, it can be called, if not, it throws the appropriate error. Fixes #133.