Fix issue starting`ASWebAuthenticationSession` on iOS 13+
Starting the `ASWebAuthenticationSession` on iOS 13 was triggering an error. Upon further investigation, it appears Apple has added a new protocol/property to `ASWebAuthenticationSession` called `presentationContextProvider` which is required to start the session.

This commit fixes the issue by providing a default implementation of the `ASWebAuthenticationPresentationContextProviding` protocol on all view controllers, and setting the `presentationContextProvider` property to the `authorizeContext`.