Fix implementation of `ASWebAuthenticationPresentationContextProviding` protocol
This protocol expects an ASPresentationAnchor to be returned (which is a UIWindow on iOS, or NSWindow on Mac OS). This commit updates the implementation to return the window object associated with the view of the `authorizeContext` (and assumes the authorize context is an instance of UIViewController).

This change is in line with Apple’s example documentation:
https://developer.apple.com/documentation/authenticationservices/authenticating_a_user_through_a_web_service